Design change: use window.open, instead of window.location.href
This was a mistake, due to a misunderstanding of the behaviour of
packaged app when a browser window was opened.  It was believed
that the packaged window address was lost -- hence, using href was
making it faster.  But window.open() lets the opener still alive!

This change makes the code simplier:
- whatever the app is packaged or not, the process is exactly the same:
  in the jaxogram window, wait for the MemoryCache to be filled with
  the verifier value obtained on the oauth browser window side;
- the problem with hosted jaxogram being re-entered (showing install,
  etc) is gone
- no more needs for referrers...
- added to the fact that the "wait" is made on the GAE (server) side,
  there are no more getVerifier() going forth and back, and that also
  helped.

Brief, it looks like the decision to go for location.href was a bad one.
It triggered about 2 days of work, which could have been avoided.