Start implementing Tweets to be associated with image uploads.
Uploading a photo (on Twitter) uses an OAuth'ed POST. I jumped on the
occasion of implementing it to simplify the oauth library (which, IMHO,
was deeply polluted by the Orkut people.)
To associate a tweet to photo upload requires to better inquire of the
tweet profile of the user.  This is not absolutely necessary, but it
leads to a nicer GUI.  But, then, it's no more an OAuth POST: it's GET.
So I had to redo for GET what I did for POST: cleaning the mess.