Lot of changes to avoid id to have meanings.
Primary goal was to increase the difficulty for reverse-engineering,
removing any human meaning to the "id" attribute. Then at examining
other attributes (such as "role") I went to the conclusion those were
misused.
After cleaning, it became obvious that, in no case, the "id" attribute
must carry a meaning, the real meaning being provided by other attributes,
or the tag name, or the position of the element in the document structure.
Even worst, several "id meanings" were either wrong, or error-prone.

This revision results from a full revamp of attributes, CSS and selectors.