Prefix all jaxo-owned strings with z_
Jumping on the occasion of "uglyfying", I decided to make clear what
strings are "jaxo owned".  That is, the string the ids of which can be
obfuscated.  Mainly, these are i18n'able strings...
Doing so, I made sure to avoid the namespace collision between
- class names
- object names (JSON)
- DOM / HTML5 / CSS3 reserved names (those ones have also to be declared
to the Google Closure Compiler, in case it is not at the latest of W3C
innovations)