"drawer" becomes "principal" and role fixes.
Going further into semantics, "drawer" expresses the how, and that should
be left to CSS.  "principal" is more appropriate.  Yes, we still have the
"sidebar" class, but it is legacy, "sidebar" meaning is something that plays
the complementary to what is principal.

Roles.  Today, I feel that roles (ARIA) is a beast I should avoid.
role='main' means nothing to the semantics.
"role" is a meaningless word, like "type", "kind", etc.
There are several "role" attributes still in there, but it's less severe.
For instance, the "listbox" role should be understood as "a container of
listed items". It's OK for nowadays.