Fixed the problem for not being able to navigate from the splash screen.
I have spent about 5 hours on it.  I tried all of the CORS features, the
HTML 5 sandbox attribute, the mozbrowser... until I found it.

The AD is displayed in an iframe. The src of the iframe is "inline" so
to avoid x-domain. To display an ad, I append to Jaxogram main HTML
something like:

  <iframe
  src="data:text/html;charset=utf-8,<html><head> ..
     <base target='_blank'/>
     <body>
     <script ...
        src='http://ad-tag.inner-active.mobi/simpleM2M/RequestTagAd?
        params=lotOfStuff>
     </script></body></html>">

This append does a request to inner-active/../RequestTagAd that returns
an HTTP snippet, so that we now have:
  <iframe src="data:text/html;charset=utf-8,<htm.><head>..<base
      target='_blank'/>..</html>">
  <html>
  <body onclick="location.reload();">
  <a href="http://ad-tag.inner-active/.../whoeverAdvertizes"><img
       src="niceImageOfWhoeverAdvertizes"/></a>
  </body></html>

We have a nice splash.  If we click on it... we are supposed to enter
the whoeverAdvertizes site.  But nothing happens. Reason is that base
target="_blank" violates security.  iframe must stay in iframe, so I
did: base target="_self". Nothing happens, again, starting driving me
nuts.  Trying all kind of tricks...

Then 5 hours later, I restarted from scratch and was lucky to look at
the code inside the iframe.

I found this "onclick=reload".

Hmmm.. the click bubbles up from the <A> to the <BODY>, and -- depending
on the way browsers assign the location.href -- the new location can be
either 1) the href'ed by A; 2) the current one reloaded.

So I did:

  <iframe
   src="data:text/html;charset=utf-8,<html><head> ..
      <base target='_self'/>
      <body>
      <script ...
         src='http://ad-tag.inner-active.mobi/simpleM2M/RequestTagAd?
         params=lotOfStuff>
      </script>
      <script language='javascript'>
      document.body.removeAttribute('onclick');
      </script></body></html>"
  >

And that worked.