The long awaited ad feature, now appearing to work on the device.
The strategy to get the ads was changed to go thru our appspot server.
As explained before, inner-active allows it if you decipher their doc.
This version uses a "custom" 15.jaxogram.appspot, but their is no real
reasons for.

Previous version of iaAd.js should be kept as a reference.