Last change for Amazon Web App try, before a potential review of this solution in case Amazon (and Chrome Webview) does the required changes.
This revision is in an unstable state. I was able to get most of the UI
working, but had to comment out some...

As seen on the Chromium branch, on which I worked from Jun 11 to june 23 2013,
SVG is a problem.  Recall that the ultimate show-stopper on this defunct thread
was that we couldn't send the filtered image.  A CSP violation was generated,
while the ISO standard allows the action.  Nowadays, I think it might be
perhaps circumvented using srcdoc (?).

New problems:
- browser windows do not open (window.open) if this action does not
  result from a user click (so I had to add an extra message)
- there is no way to know when the App window is again active
  (hence, the polling for OAuth response could get infinite).
  Although, there might be other means...

But I didn't pursue, because of this:
| Amazon's HTML5 runtime supports most major features in the standard
| except for features that are not supported by the Kindle Fire, such
| as hardware components like GPS, camera access, accelerometer and or
| the microphone.

Either you are HTML5 or you are not, Amazonians.
And, definitively, you are not.